Update hooks and deployment scripts

+++ contracts/libraries/MessageSender.sol
@@ -22,20 +22,14 @@ library MessageSender {
         address receiver,
         Client.EVMTokenAmount[] memory tokensToSendDetails,
         PayFeesIn payFeesIn,
-        uint _gasLimit
-    ) internal returns(bytes32) {
+        uint256 _gasLimit
+    ) internal returns (bytes32) {
         uint256 length = tokensToSendDetails.length;
-        require(
-            length <= i_maxTokensLength,
-            "Maximum 5 different tokens can be sent per CCIP Message"
-        );
+        require(length <= i_maxTokensLength, "Maximum 5 different tokens can be sent per CCIP Message");
 
-        for (uint256 i = 0; i < length; ) {
+        for (uint256 i = 0; i < length;) {
             if (tokensToSendDetails[i].token != i_link) {
-                IERC20(tokensToSendDetails[i].token).approve(
-                    i_router,
-                    tokensToSendDetails[i].amount
-                );
+                IERC20(tokensToSendDetails[i].token).approve(i_router, tokensToSendDetails[i].amount);
             }
             unchecked {
                 ++i;
@@ -53,23 +47,14 @@ library MessageSender {
             feeToken: payFeesIn == PayFeesIn.LINK ? i_link : address(0)
         });
 
-        uint256 fee = IRouterClient(i_router).getFee(
-            destinationChainSelector,
-            message
-        );
+        uint256 fee = IRouterClient(i_router).getFee(destinationChainSelector, message);
 
         bytes32 messageId;
 
         if (payFeesIn == PayFeesIn.LINK) {
-            messageId = IRouterClient(i_router).ccipSend(
-                destinationChainSelector,
-                message
-            );
+            messageId = IRouterClient(i_router).ccipSend(destinationChainSelector, message);
         } else {
-            messageId = IRouterClient(i_router).ccipSend{value: fee}(
-                destinationChainSelector,
-                message
-            );
+            messageId = IRouterClient(i_router).ccipSend{value: fee}(destinationChainSelector, message);
         }
 
         // emit IndexFactory.MessageSent(messageId);
@@ -89,8 +74,8 @@ library MessageSender {
         address receiver,
         bytes memory _data,
         PayFeesIn payFeesIn,
-        uint _gasLimit
-    ) internal returns(bytes32) {
+        uint256 _gasLimit
+    ) internal returns (bytes32) {
         require(i_router != address(0), "Invalid i_router address");
         require(i_link != address(0), "Invalid i_link address");
         require(destinationChainSelector != 0, "Invalid destinationChainSelector");
@@ -105,26 +90,16 @@ library MessageSender {
                 Client.EVMExtraArgsV1({gasLimit: _gasLimit})
             ),
             feeToken: payFeesIn == PayFeesIn.LINK ? i_link : address(0)
-
         });
 
-        uint256 fee = IRouterClient(i_router).getFee(
-            destinationChainSelector,
-            message
-        );
+        uint256 fee = IRouterClient(i_router).getFee(destinationChainSelector, message);
 
         bytes32 messageId;
 
         if (payFeesIn == PayFeesIn.LINK) {
-            messageId = IRouterClient(i_router).ccipSend(
-                destinationChainSelector,
-                message
-            );
+            messageId = IRouterClient(i_router).ccipSend(destinationChainSelector, message);
         } else {
-            messageId = IRouterClient(i_router).ccipSend{value: fee}(
-                destinationChainSelector,
-                message
-            );
+            messageId = IRouterClient(i_router).ccipSend{value: fee}(destinationChainSelector, message);
         }
 
         // emit IndexFactory.MessageSent(messageId);
